MyPharmacyApp V4

+++ composeApp/src/commonMain/kotlin/data/entity/Medicine.kt
@@ -3,9 +3,9 @@ package data.entity
 import kotlinx.datetime.LocalDate
 
 data class Medicine(
-    val id: Int? = null,
+    val id: Long? = null,
     val name: String,
     val description: String?,
     val expirationDate: LocalDate?,
-    val typeId: Int?
+    val typeId: Long?
 )
